Fix Nix build
The older version of `zig-overlay` used was unable to handle the newer
directory layout of Zig tarballs, so `zig-overlay` has been updated and
is currently using a `master` version of Zig.

Additionally, `zls.json` was removed since `zls` will now only look for
it in local or global configuration folders.

Closes #330.