Add known-folders as a flake input
This allows the flake to work without passing in `?submodules=1`, which
makes it easy to include in other flakes.

This commit also makes it possible to override the path to the
`known-folders` package like so:

```
zig build -Dknown-folders=/path/to/known-folders.zig
```

This allows `flake.nix` to pass in the nix store path.