Search for r_brace at the AST level, not the source code level.
Before we were doing a search for the right brace at the source code level,
but only in the handler for the block scope. This was fallible, since it only
searched up to 16 characters from the l_brace. Another problem was that this
search was not being done for any parent function scopes. So the parent scope
would end up being shorter than its child block, causing any declaration
searches within the block to terminate early.