simplify `offsets.sourceIndexToTokenIndex`
This function contains a special case when the number of tokens is
under 600 which manually unrolls the binary search iterations.

I don't see any need for an overengineered implementation of binary search on tiny source files.

Note that I've doubled the threshold because the old version did the check after doing one iteration.