make the Zig+ZLS compatibility check more strict
The compatibility check was previously only executed on the config
resolve at startup/initialize. This check has moved to the
updateConfiguration function which also includes updates that happen 
through in-editor configuration.

ZLS will also warn when the minimum Zig version is not met at runtime.
The only exception is when using Zig `0.11.*` or `0.10.*` (no development builds).