Send $/progress messages on build file parsing
Initial build processing can be slow. It can be desirable to know when
the initial processing is finished to ensure that LSP queries are
processing the entire code base

Implement LSP protocol for this $/progress. Message flow is as follows

* Client publishes that it supports the window/workDoneProgress/create
  command on initialize
* Something triggers a build file invalidation
* Server requests that the client creates a work done progress report
* Server sends the initial work done progress
* Server sends the final work done progress

Note that this commit does not attempt to send intermediate reports.
AFAICT this would involve some fairly complex hooking of the build
runner that probably isn't worth it

Implementation details:
* DocumentStore owns the build file invalidation, but Server owns the
  message serialization/sending. In order to support DocumentStore
  triggering Server actions, we provide a WorkerNotifier that the Server
  populates with the logic needed to send messages to the client
* In some cases multiple builds can happen in parallel. We need to
  ensure we aren't double starting the same progress message. Fold them
  all builds into one