Deprecate 'semantic_tokens' config in favor of 'augmentsSyntaxTokens'
The `augmentsSyntaxTokens` client capability can be used to indicate
that the client is going to combine existing syntax tokens with the
semantic tokens of the server. This can not only reduce the amount of
data that is transferred over the protocol but can also lead to better
highlighting as syntax tokens often more accurately distinguish between
different language constructs on the syntax level. An example would be
keywords where the LSP protocol has a single standard token type
compared to syntax highlighting where different keywords are often
differentiated between based on their purpose.

The 'semantic_tokens' config had been added for this purpose. The
'partial' option is used to prevent LSP semantic tokens to override
existing syntax highlighting of the editor. The benefit of the client
capability is that this decision is automatically made by supported
editors instead of relying on user to select the right configuration.